initial impl of realtime deleted doc filter in completion
