+++ src/main/java/org/elasticsearch/search/suggest/completion/Completion090PostingsFormat.java
@@ -27,18 +27,13 @@ import org.apache.lucene.search.suggest.Lookup;
 import org.apache.lucene.store.IOContext.Context;
 import org.apache.lucene.store.IndexInput;
 import org.apache.lucene.store.IndexOutput;
-import org.apache.lucene.store.InputStreamDataInput;
-import org.apache.lucene.store.OutputStreamDataOutput;
 import org.apache.lucene.util.BytesRef;
 import org.apache.lucene.util.IOUtils;
 import org.elasticsearch.ElasticsearchIllegalStateException;
 import org.elasticsearch.common.logging.ESLogger;
 import org.elasticsearch.common.logging.Loggers;
 import org.elasticsearch.index.mapper.core.CompletionFieldMapper;
-import org.elasticsearch.search.suggest.completion.CompletionTokenStream.ToFiniteStrings;
 
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.util.Comparator;
 import java.util.Iterator;
@@ -302,56 +297,6 @@ public class Completion090PostingsFormat extends PostingsFormat {
         }
     }
 
-    public static abstract class CompletionLookupProvider implements PayloadProcessor, ToFiniteStrings {
-
-        public static final char UNIT_SEPARATOR = '\u001f';
-
-        public abstract FieldsConsumer consumer(IndexOutput output) throws IOException;
-
-        public abstract String getName();
-
-        public abstract LookupFactory load(IndexInput input) throws IOException;
-
-        @Override
-        public BytesRef buildPayload(BytesRef surfaceForm, long weight, BytesRef payload) throws IOException {
-            if (weight < -1 || weight > Integer.MAX_VALUE) {
-                throw new IllegalArgumentException("weight must be >= -1 && <= Integer.MAX_VALUE");
-            }
-            for (int i = 0; i < surfaceForm.length; i++) {
-                if (surfaceForm.bytes[i] == UNIT_SEPARATOR) {
-                    throw new IllegalArgumentException(
-                            "surface form cannot contain unit separator character U+001F; this character is reserved");
-                }
-            }
-            ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
-            OutputStreamDataOutput output = new OutputStreamDataOutput(byteArrayOutputStream);
-            output.writeVLong(weight + 1);
-            output.writeVInt(surfaceForm.length);
-            output.writeBytes(surfaceForm.bytes, surfaceForm.offset, surfaceForm.length);
-            output.writeVInt(payload.length);
-            output.writeBytes(payload.bytes, 0, payload.length);
-
-            output.close();
-            return new BytesRef(byteArrayOutputStream.toByteArray());
-        }
-
-        @Override
-        public void parsePayload(BytesRef payload, SuggestPayload ref) throws IOException {
-            ByteArrayInputStream byteArrayInputStream = new ByteArrayInputStream(payload.bytes, payload.offset, payload.length);
-            InputStreamDataInput input = new InputStreamDataInput(byteArrayInputStream);
-            ref.weight = input.readVLong() - 1;
-            int len = input.readVInt();
-            ref.surfaceForm.grow(len);
-            ref.surfaceForm.length = len;
-            input.readBytes(ref.surfaceForm.bytes, ref.surfaceForm.offset, ref.surfaceForm.length);
-            len = input.readVInt();
-            ref.payload.grow(len);
-            ref.payload.length = len;
-            input.readBytes(ref.payload.bytes, ref.payload.offset, ref.payload.length);
-            input.close();
-        }
-    }
-
     public CompletionStats completionStats(IndexReader indexReader, String ... fields) {
         CompletionStats completionStats = new CompletionStats();
         for (AtomicReaderContext atomicReaderContext : indexReader.leaves()) {
@@ -372,10 +317,4 @@ public class Completion090PostingsFormat extends PostingsFormat {
         return completionStats;
     }
 
-    public static abstract class LookupFactory {
-        public abstract Lookup getLookup(CompletionFieldMapper mapper, CompletionSuggestionContext suggestionContext);
-        public abstract CompletionStats stats(String ... fields);
-        abstract AnalyzingCompletionLookupProvider.AnalyzingSuggestHolder getAnalyzingSuggestHolder(CompletionFieldMapper mapper);
-        public abstract long ramBytesUsed();
-    }
 }
